+++ wp-content/plugins/wp-rabbit/lib/varnish.php
@@ -23,36 +23,7 @@ add_action( 'save_post', function( $post_id ) {
     return;
   }
 
-  $parse = parse_url($post_url);
-
-  // The x-access-token is arbitrary.
-  $purge_request_args = array(
-    "method" => "PURGE",
-    "headers" => array(
-      "X-Access-Token" => isset( $_SERVER['HTTP_X_SELECTED_CONTAINER'] ) ? $_SERVER['HTTP_X_SELECTED_CONTAINER'] : '',
-      "Host" => $parse['host'],
-      "X-Set-Branch" => isset( $_SERVER['GIT_BRANCH'] ) ? $_SERVER['GIT_BRANCH'] : ''
-    ),
-    "blocking" => defined( "WP_DEBUG" ) && WP_DEBUG ? true : false
-  );
-
-  // make purge request to wpcloud.io. (this gets public DNS of wpcloud servers and then uses GCE load balancers to purge the appropriate machien based on hostname)
-  $_purge = wp_remote_request( "http://c.rabbit.ci" . $parse['path'], $purge_request_args );
-
-  rabbit_write_log( 'Post [' . $post_id. '] updated. Purging cache at [' . $parse['host'] . $parse['path'] . '] url for ['.$_SERVER['GIT_BRANCH'].'] branch.' );
-
-  if( $_purge && wp_remote_retrieve_body( $_purge ) && defined( "WP_DEBUG" ) ) {
-
-    try {
-      $_response = json_decode( wp_remote_retrieve_body( $_purge ) );
-      rabbit_write_log( ' - Purge response [' . $_response->message . '].' );
-
-    } catch ( Exception $e ) {
-      rabbit_write_log( ' - Unable to parse purge response.' );
-    }
-
-  }
-
+  rabbit_purge_url( $post_url );
 
 }, 50 );
 
@@ -64,24 +35,7 @@ add_action( 'save_post', function( $post_id ) {
 add_filter( 'generate_rewrite_rules', function( $wp_rewrite ) {
 
   // disabled for now, seems to run all the time on UD, for instance.
-  return $wp_rewrite->rules;
-
-  $site_url = home_url();
-
-  $parse = parse_url($site_url);
-
-  // make purge request to wpcloud.io for entire sute
-  wp_remote_request( "http://wpcloud.io" . $parse['path'] . '/*', array(
-    "method" => "PURGE",
-    "headers" => array(
-      "X-Access-Token" => "yhcwokwserjzdjir",
-      "Host" => $parse['host'],
-      "x-set-branch" => isset( $_SERVER['GIT_BRANCH'] ) ? $_SERVER['GIT_BRANCH'] : ''
-    ),
-    "blocking" => defined( "WP_DEBUG" ) && WP_DEBUG ? true : false
-  ));
-
-  rabbit_write_log( 'Purging all cache for [' . $parse['host'] . $parse['path'] . '/*' . '] in ['.$_SERVER['GIT_BRANCH'].'] branch.' );
+  // rabbit_purge_url( home_url( '/*' ) );
 
   return $wp_rewrite->rules;
 
@@ -101,3 +55,87 @@ function rabbit_write_log( $data ) {
   }
 
 }
+
+/**
+ * Purge URL in Varnish/CloudFront. This is a nob-blocking request by default.
+ *
+ * This method will automatically utilize the current branch.
+ *
+ * Purge all cache:
+ *
+ *    rabbit_purge_url( home_url('/*') );
+ *
+ * Purge custom URL
+ *    rabbit_purge_url( home_url('/my-custom-url') );
+ *
+ * Purge custom URL with custom get arguments.
+ *    rabbit_purge_url( home_url('/my-custom-url?test=one') );
+ *
+ * Purge a permalink post.
+ *    rabbit_purge_url( get_permalink('123456') );
+ *
+ * All purge activity written to /var/log/wpcloud.site/deployment.log file.
+ *
+ * @param $url
+ * @param array $args
+ * @return bool
+ */
+function rabbit_purge_url( $url, $args = array() ) {
+
+  $args = wp_parse_args($args, array(
+    'blocking' => defined( "WP_DEBUG" ) && WP_DEBUG ? true : false
+  ));
+
+  if( !$url ) {
+    return false;
+  }
+
+  $parse = parse_url($url);
+
+  if( !$parse['path'] ) {
+    return false;
+  }
+
+  // The x-access-token is arbitrary.
+  $purge_request_args = array(
+    "method" => "PURGE",
+    "headers" => array(
+      "X-Access-Token" => isset( $_SERVER['HTTP_X_SELECTED_CONTAINER'] ) ? $_SERVER['HTTP_X_SELECTED_CONTAINER'] : '',
+      "Host" => $parse['host'],
+      "X-Set-Branch" => isset( $_SERVER['GIT_BRANCH'] ) ? $_SERVER['GIT_BRANCH'] : ''
+    ),
+    "blocking" => isset( $args['blocking'] ) ? $args['blocking'] : false
+  );
+
+  $_purge_url =  "http://c.rabbit.ci" . $parse['path'];
+
+  if( isset( $parse['query'] ) ) {
+    $_purge_url = $_purge_url . '?' . $parse['query'];
+  }
+
+  // make purge request to wpcloud.io. (this gets public DNS of wpcloud servers and then uses GCE load balancers to purge the appropriate machien based on hostname)
+  $_purge = wp_remote_request( $_purge_url, $purge_request_args );
+
+  if( isset( $args['post_id'] ) ) {
+    rabbit_write_log( 'Post [' . $args['post_id']. '] updated. Purging cache at [' . $parse['host'] . $_purge_url . '] url for ['.$_SERVER['GIT_BRANCH'].'] branch.' );
+  } else {
+    rabbit_write_log( 'Purging cache at [' . $parse['host'] . $_purge_url . '] url for ['.$_SERVER['GIT_BRANCH'].'] branch.' );
+  }
+
+  if( $_purge && wp_remote_retrieve_body( $_purge ) && defined( "WP_DEBUG" ) ) {
+
+    try {
+      $_response = json_decode( wp_remote_retrieve_body( $_purge ) );
+      rabbit_write_log( ' - Purge response for [' . $_purge_url . '], [' . $_response->message . '].' );
+
+    } catch ( Exception $e ) {
+      rabbit_write_log( ' - Unable to parse purge response.' );
+    }
+
+  }
+
+  return isset( $_response ) ? $_response : true;
+
+}
+
+