+++ wp-content/plugins/wp-rets-client/lib/classes/class-xmlrpc.php
@@ -45,6 +45,74 @@ namespace UsabilityDynamics\WPRETSC {
         // REST API
         add_action( 'rest_api_init', array( $this, 'api_init' ), 100 );
 
+        // Run actions after a property is published. (e.g. associated Agents)
+        add_action( 'wrc_property_published', array( $this, 'property_published' ), 100, 2 );
+
+      }
+
+      /**
+       * Handle post-publish actions.
+       *
+       * - Associate wpp_agents based on wpp_agency_agent terms.
+       * - Agent Users must have a "rets_id" field that matches their unique ID in the MLS.
+       * - the rets.ci service adds agents/office as terms with extra meta to store information such as their MLS/MLN.
+       *
+       * @author potanin@UD
+       */
+      public function property_published( $post_id, $post_data ) {
+        // ud_get_wp_rets_client()->write_log( "Running [property_published] for [" . $post_id . "].", 'debug' );
+
+        // If missing agent module, do nothing.
+        if( !function_exists( 'ud_get_wpp_agents' ) ) {
+          return;
+        }
+
+        $_wpp_agents = array();
+
+        // get "agent" terms.
+        $_agents = wp_get_post_terms( $post_id, 'wpp_agency_agent', array(
+          'orderby'    => 'count',
+          'hide_empty' => false,
+          'fields' => 'ids'
+        ) );
+
+        if( $_agents && is_array( $_agents ) && count($_agents) >= 1 ) {
+
+          foreach( $_agents as $_agent_term_id ) {
+
+            // get the "rets_id" field from agent term.
+            $_rets_id = get_term_meta( $_agent_term_id, 'listing-agent-rets_id', true );
+
+            // Find relevant WP agent users with matching "rets_id" field.
+            $_user_agents = get_users(array(
+              'meta_key' => 'rets_id',
+              'meta_value' => $_rets_id,
+              'fields' => array( 'ID', 'display_name' ),
+              'number' => 1
+            ));
+
+            if( is_array( $_user_agents ) ) {
+
+              foreach( $_user_agents as $_user_object ) {
+                $_wpp_agents[] = $_user_object->ID;
+                ud_get_wp_rets_client()->write_log( "Found agent-user [" . $_user_object->display_name . "] with rets_id [" . $_rets_id . "] for listing [" . $post_id . "].", 'info' );
+              }
+
+            }
+
+          }
+
+          // remove all past agent(s)
+          delete_post_meta( $post_id, 'wpp_agents' );
+
+          // add new agents, if the exist.
+          foreach( (array) $_wpp_agents as $_agent_user_id ){
+            add_post_meta($post_id, 'wpp_agents', $_agent_user_id );
+            // ud_get_wp_rets_client()->write_log( "Associated agent [" . $_agent_user_id . "] to listing [" . $post_id . "].", 'info' );
+          }
+
+
+        }
 
       }
 
@@ -741,7 +809,10 @@ namespace UsabilityDynamics\WPRETSC {
         $options = wp_parse_args( isset( $post_data['_options'] ) ? $post_data['_options'] : array(), array(
           'skipTermCounting' => false,
           'skipTermUpdates' => false,
-          'skipMediaUpdate' => false
+          'skipMediaUpdate' => false,
+          'skipSlideshowImages' => true,
+          'createWPPAttributes' => false,
+          'createWPPTerms' => false
         ));
 
         ud_get_wp_rets_client()->write_log( 'Have request [wpp.createProperty] request.', 'debug' );
@@ -813,11 +884,26 @@ namespace UsabilityDynamics\WPRETSC {
         // Insert all the terms and creates taxonomies.
         if( !isset( $options[ 'skipTermUpdates' ] ) || !$options[ 'skipTermUpdates' ] ) {
           Utility::insert_property_terms( $_post_id, $_post_data_tax_input, $post_data );
+          if( isset( $options[ 'createWPPTerms' ] ) && $options[ 'createWPPTerms' ] ) {
+            Utility::create_wpp_taxonomies( array_keys( (array)$_post_data_tax_input ) );
+          }
         }
 
         if( !isset( $options[ 'skipMediaUpdate' ] ) || !$options[ 'skipMediaUpdate' ] ) {
           Utility::insert_media( $_post_id, $post_data[ '_media' ] );
         }
+
+        if( !isset( $options[ 'skipSlideshowImages' ] ) || !$options[ 'skipSlideshowImages' ] ) {
+          Utility::insert_slideshow_images( $_post_id );
+        }
+
+        $postmeta = array();
+        foreach( (array) $post_data[ 'meta_input' ] as $_meta_key => $_meta_value ) {
+          if( !empty( $_meta_value ) && isset( $options[ 'createWPPAttributes' ] ) && $options[ 'createWPPAttributes' ] ) {
+            array_push( $postmeta, $_meta_key );
+          }
+        }
+        Utility::create_wpp_attributes( $postmeta );
 
         if( $_post_id ) {
           ud_get_wp_rets_client()->write_log( 'Updating property post [' . $_post_id  . '].', 'debug' );
@@ -858,7 +944,7 @@ namespace UsabilityDynamics\WPRETSC {
           /**
            * Do something after property is published
            */
-          do_action( 'wrc_property_published', $_post_id );
+          do_action( 'wrc_property_published', $_post_id, $post_data );
 
         } else {
           ud_get_wp_rets_client()->write_log( 'Error publishing post ' . $_post_id, 'error' );
@@ -874,6 +960,8 @@ namespace UsabilityDynamics\WPRETSC {
         }
 
         ud_get_wp_rets_client()->write_log( 'Term counting complete for [' . $_post_id . '].', 'info' );
+
+        self::flush_cache( $_post_id );
 
         return array(
           "ok" => true,
@@ -904,7 +992,10 @@ namespace UsabilityDynamics\WPRETSC {
         $options = wp_parse_args( isset( $post_data['_options'] ) ? $post_data['_options'] : array(), array(
           'skipTermCounting' => false,
           'skipTermUpdates' => false,
-          'skipMediaUpdate' => false
+          'skipMediaUpdate' => false,
+          'skipSlideshowImages' => true,
+          'createWPPAttributes' => false,
+          'createWPPTerms' => false
         ));
 
         ud_get_wp_rets_client()->write_log( 'Have request [wpp.updateProperty] request.', 'debug' );
@@ -934,18 +1025,26 @@ namespace UsabilityDynamics\WPRETSC {
           $wpdb->update( $wpdb->posts, array( 'post_content' => $post_data[ 'post_content' ] ), array( 'ID' => $post_data['ID' ] ) );
         }
 
+        $postmeta = array();
         foreach( (array) $post_data[ 'meta_input' ] as $_meta_key => $_meta_value ) {
           update_post_meta( $post_data['ID' ], $_meta_key, $_meta_value );
+          if( !empty( $_meta_value ) && isset( $options[ 'createWPPAttributes' ] ) && $options[ 'createWPPAttributes' ] ) {
+            array_push($postmeta, $_meta_key);
+          }
         }
+        Utility::create_wpp_attributes( $postmeta );
 
         if( (isset( $options[ 'skipTermUpdates' ] ) || !$options[ 'skipTermUpdates' ]) && isset($post_data[ 'tax_input' ]) ) {
           Utility::insert_property_terms( $post_data[ 'ID' ], $post_data[ 'tax_input' ], $post_data );
+          if( isset( $options[ 'createWPPTerms' ] ) && $options[ 'createWPPTerms' ] ) {
+            Utility::create_wpp_taxonomies( array_keys( (array)$post_data[ 'tax_input' ] ) );
+          }
           ud_get_wp_rets_client()->write_log( 'Updated terms.', 'debug' );
         }
 
         ud_get_wp_rets_client()->write_log( 'Property update finished, clearing cache.', 'debug' );
 
-        clean_post_cache( $post_data[ 'ID' ] );
+        self::flush_cache( $post_data[ 'ID' ] );
 
         return array(
           "ok" => true,
@@ -979,7 +1078,10 @@ namespace UsabilityDynamics\WPRETSC {
         $options = wp_parse_args( isset( $post_data['_options'] ) ? $post_data['_options'] : array(), array(
           'skipTermCounting' => false,
           'skipTermUpdates' => false,
-          'skipMediaUpdate' => false
+          'skipMediaUpdate' => false,
+          'skipSlideshowImages' => true,
+          'createWPPAttributes' => false,
+          'createWPPTerms' => false
         ));
 
         // Defer term counting until method called again.
@@ -1046,11 +1148,26 @@ namespace UsabilityDynamics\WPRETSC {
         // Insert all the terms and creates taxonomies.
         if( !isset( $options[ 'skipTermUpdates' ] ) || !$options[ 'skipTermUpdates' ] ) {
           Utility::insert_property_terms( $_post_id, $_post_data_tax_input, $post_data );
+          if( isset( $options[ 'createWPPTerms' ] ) && $options[ 'createWPPTerms' ] ) {
+            Utility::create_wpp_taxonomies( array_keys( (array)$_post_data_tax_input ) );
+          }
         }
 
         if( !isset( $options[ 'skipMediaUpdate' ] ) || !$options[ 'skipMediaUpdate' ] ) {
           Utility::insert_media( $_post_id, $post_data[ '_media' ] );
         }
+
+        if( !isset( $options[ 'skipSlideshowImages' ] ) || !$options[ 'skipSlideshowImages' ] ) {
+          Utility::insert_slideshow_images( $_post_id );
+        }
+
+        $postmeta = array();
+        foreach( (array) $post_data[ 'meta_input' ] as $_meta_key => $_meta_value ) {
+          if( !empty( $_meta_value ) && isset( $options[ 'createWPPAttributes' ] ) && $options[ 'createWPPAttributes' ] ) {
+            array_push( $postmeta, $_meta_key );
+          }
+        }
+        Utility::create_wpp_attributes( $postmeta );
 
         if( $_post_id ) {
           ud_get_wp_rets_client()->write_log( 'Updating property post [' . $_post_id  . '].', 'debug' );
@@ -1091,7 +1208,7 @@ namespace UsabilityDynamics\WPRETSC {
           /**
            * Do something after property is published
            */
-          do_action( 'wrc_property_published', $_post_id );
+          do_action( 'wrc_property_published', $_post_id, $post_data );
 
         } else {
           ud_get_wp_rets_client()->write_log( 'Error publishing post ' . $_post_id, 'error' );
@@ -1113,6 +1230,8 @@ namespace UsabilityDynamics\WPRETSC {
         );
 
         ud_get_wp_rets_client()->write_log( 'Sending [wpp.editProperty] reponse.', 'debug' );
+
+        self::flush_cache( $_post_id );
 
         return $_response;
 
@@ -1219,7 +1338,7 @@ namespace UsabilityDynamics\WPRETSC {
           ud_get_wp_rets_client()->logfile = !empty( $data[ 'logfile' ] ) ? $data[ 'logfile' ] : ud_get_wp_rets_client()->logfile;
         }
 
-        ud_get_wp_rets_client()->write_log( 'Have wpp.deleteProperty request.', 'info' );
+        ud_get_wp_rets_client()->write_log( 'Have request [wpp.deleteProperty].', 'info' );
 
         if( !$post_id || !is_numeric( $post_id ) ) {
           ud_get_wp_rets_client()->write_log(  'No post ID provided', 'info' );
@@ -1267,6 +1386,8 @@ namespace UsabilityDynamics\WPRETSC {
 
         $response['time' ] = timer_stop();
 
+        self::flush_cache( $post_id );
+
         return $response;
 
       }
@@ -1302,7 +1423,7 @@ namespace UsabilityDynamics\WPRETSC {
           ud_get_wp_rets_client()->logfile = !empty( $data[ 'logfile' ] ) ? $data[ 'logfile' ] : ud_get_wp_rets_client()->logfile;
         }
 
-        ud_get_wp_rets_client()->write_log( 'Have [wpp.trashProperty] request. Post id: ' . $post_id, 'info' );
+        ud_get_wp_rets_client()->write_log( 'Have request [wpp.trashProperty]. Post id: ' . $post_id, 'info' );
 
         if( !$post_id || !is_numeric( $post_id ) ) {
           ud_get_wp_rets_client()->write_log(  'No post ID provided', 'info' );
@@ -1310,13 +1431,19 @@ namespace UsabilityDynamics\WPRETSC {
           return $response;
         }
 
-        ud_get_wp_rets_client()->write_log( "Checking post ID [$post_id]." );
+        ud_get_wp_rets_client()->write_log( "Checking post ID [$post_id].", 'info' );
 
-        $wpdb->update( $wpdb->posts, array( 'post_status' => 'trash' ), array( 'ID' => $post_id ) );
+        // We must do 'trash' post using native function, instead of direct SQL requests....
+        // because of different bugs and issues with property status on end.... peshkov@UD
+        wp_trash_post( $post_id );
 
-        ud_get_wp_rets_client()->write_log( "Property [$post_id] trashed." );
+        ud_get_wp_rets_client()->write_log( $wpdb->last_error, 'info' );
+
+        ud_get_wp_rets_client()->write_log( "Property [$post_id] trashed.", 'info' );
 
         $response['time' ] = timer_stop();
+
+        self::flush_cache( $post_id );
 
         return $response;
 
@@ -1693,6 +1820,17 @@ namespace UsabilityDynamics\WPRETSC {
 
         die( json_encode( $data, JSON_PRETTY_PRINT ) );
 
+      }
+
+      /**
+       * Flush Object Cache for particular property
+       *
+       * @param int $post_id
+       */
+      static protected function flush_cache( $post_id ) {
+        ud_get_wp_rets_client()->write_log( "Flushing object cache for [" . $post_id . "] post_id", 'info' );
+        clean_post_cache( $post_id );
+        do_action( 'wprc::xmlrpc::on_flush_cache', $post_id );
       }
 
     }