+++ wp-content/plugins/wp-rabbit/changes.md
@@ -1,5 +1,6 @@
 ### 1.0.8
 * Admin navbar now displays host/port used by Redis/Memcached servers, if Object Cache is used.
+* Added `rabbit_purge_url` method for triggering URL purging as neeeded.
 
 ### 1.0.7
 * Performance improvement to self-updater logic.